Shells/helm-service-v2/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
GitPython==3.1.14
requests==2.26.0